transport-helper: ensure helper_data.name is also freed
Both invocations of transport_helper_init take ownership of name, see:
- Path 1:
  helper = xstrndup(...)
    https://git.kernel.org/pub/scm/git/git.git/tree/transport.c#n1049
  followed by transport_helper_init taking ownership:
    https://git.kernel.org/pub/scm/git/git.git/tree/transport.c#n1053
- Path 2:
  handler = xmemdupz(...), immediately followed by transpot_helper_init
  taking ownership:
    https://git.kernel.org/pub/scm/git/git.git/tree/transport.c#n1083

Therefore it seems clear that all functions clearing helper_data
should also be responsible for clearing the name. (We also add
a comment to transport_helper_init to make the ownership explicit.)

This avoids leaks such as the following, which I used to see on
every git pull - it's minor, but noisy. In most cases there appears
to be only one transport (so we'd leak at most once). The biggest
exception is builtin/push.c which can create multiple transports
in a loop, but even there the impact is unlikely to be noticeable.

Direct leak of 6 byte(s) in 1 object(s) allocated from:
    #0 0x49a53d in malloc /home/abuild/rpmbuild/BUILD/llvm-11.0.0.src/build/../projects/compiler-rt/lib/asan/asan_malloc_linux.cpp:145:3
    #1 0x9a6fe4 in do_xmalloc /home/ahunt/oss-fuzz/git/wrapper.c:41:8
    #2 0x9a7078 in do_xmallocz /home/ahunt/oss-fuzz/git/wrapper.c:75:8
    #3 0x9a705a in xmallocz /home/ahunt/oss-fuzz/git/wrapper.c:83:9
    #4 0x9a7107 in xmemdupz /home/ahunt/oss-fuzz/git/wrapper.c:99:16
    #5 0x975338 in transport_get /home/ahunt/oss-fuzz/git/transport.c:1083:19
    #6 0x5636c0 in prepare_transport /home/ahunt/oss-fuzz/git/builtin/fetch.c:1388:14
    #7 0x56294d in fetch_one /home/ahunt/oss-fuzz/git/builtin/fetch.c:1837:15
    #8 0x561e2c in cmd_fetch /home/ahunt/oss-fuzz/git/builtin/fetch.c:1993:12
    #9 0x4cd60d in run_builtin /home/ahunt/oss-fuzz/git/git.c:453:11
    #10 0x4cb2da in handle_builtin /home/ahunt/oss-fuzz/git/git.c:704:3
    #11 0x4ccc37 in run_argv /home/ahunt/oss-fuzz/git/git.c:771:4
    #12 0x4cac29 in cmd_main /home/ahunt/oss-fuzz/git/git.c:902:19
    #13 0x69c43e in main /home/ahunt/oss-fuzz/git/common-main.c:52:11
    #14 0x7fec22a90349 in __libc_start_main (/lib64/libc.so.6+0x24349)

SUMMARY: AddressSanitizer: 6 byte(s) leaked in 1 allocation(s).

Signed-off-by: Andrzej Hunt <ajrhunt@google.com>